chore(suite): move getGraphDataForInterval to graph utils
